requirements.txt
httpx
markdownify
mammoth
numpy
pypdf
python-dotenv
